refactor: move `Plugins` to `plugins.rs` and add unit tests
